Do not start a new shell when changing sections
This version will not start a new shell each time you move between gam profiles.
There is a little bit more code in the gamselect.sh to set red and green PS1, 
and the alias is altered to include the source command to activate the new PS1.

As I have 0 byte json files in the .gam folder, my DEFAULT section is always safe, 
and therefore green; no gam command can be run there. I recommend you do the same.
Notice that I in the .bashrc auto-sets default section on start. Yeah, that's good.
Therefore my default PS1 colour is 32, green, and not 31, red.